+++ packages/common/src/relay/calling/actions/PlayAction.ts
@@ -1,6 +1,6 @@
 import { BaseAction } from './BaseAction'
 import { Play } from '../components/Play'
-import { PlayResult, PlayPauseResult, PlayResumeResult } from '../results'
+import { PlayResult, PlayPauseResult, PlayResumeResult, PlayVolumeResult } from '../results'
 
 export class PlayAction extends BaseAction {
 
@@ -16,11 +16,18 @@ export class PlayAction extends BaseAction {
     return this.component.stop()
   }
 
-  pause() {
-    return this.component.pause(PlayPauseResult)
+  async pause() {
+    const res = await this.component.pause()
+    return new PlayPauseResult(res)
   }
 
-  resume() {
-    return this.component.resume(PlayResumeResult)
+  async resume() {
+    const res = await this.component.resume()
+    return new PlayResumeResult(res)
+  }
+
+  async volume(value: number) {
+    const res = await this.component.volume(value)
+    return new PlayVolumeResult(res)
   }
 }