+++ packages/common/src/relay/calling/components/Controllable.ts
@@ -26,13 +26,32 @@ export abstract class Controllable extends BaseComponent {
     return new StopResult(result)
   }
 
-  async pause<T>(ResultObject: new (b: boolean) => T): Promise<T> {
+  async pause(): Promise<boolean> {
     const { code } = await this._execute(`${this.method}.pause`)
-    return new ResultObject(code === '200')
+    return code === '200'
   }
 
-  async resume<T>(ResultObject: new (b: boolean) => T): Promise<T> {
+  async resume(): Promise<boolean> {
     const { code } = await this._execute(`${this.method}.resume`)
-    return new ResultObject(code === '200')
+    return code === '200'
+  }
+
+  async volume(value: number): Promise<boolean> {
+    try {
+      const msg = new Execute({
+        protocol: this.call.relayInstance.session.relayProtocol,
+        method: `${this.method}.volume`,
+        params: {
+          node_id: this.call.nodeId,
+          call_id: this.call.id,
+          control_id: this.controlId,
+          volume: value
+        }
+      })
+      await this.call._execute(msg)
+      return true
+    } catch (error) {
+      return false
+    }
   }
 }