+++ packages/node/tests/relay/Call.test.ts
@@ -305,17 +305,17 @@ describe('Call', () => {
         { type: 'tts', params: { text: 'hello jest' } }
       ]
 
-      const getMsg = (...play: (IRelayCallingPlay | ICallingPlay)[]) => new Execute({
-        protocol: 'signalwire_service_random_uuid',
-        method: 'calling.play',
-        params: { node_id: call.nodeId, call_id: call.id, control_id: 'mocked-uuid', play }
-      })
+      const getMsg = (play: (ICallingPlay | IRelayCallingPlay)[], volume = 0) => {
+        const params: any = { node_id: call.nodeId, call_id: call.id, control_id: 'mocked-uuid', play }
+        if (volume !== 0) params.volume = volume
+        return new Execute({ protocol: 'signalwire_service_random_uuid', method: 'calling.play', params })
+      }
 
       it('.play() should wait until the playing ends', done => {
         call.play(...media).then(result => {
           expect(result).toBeInstanceOf(PlayResult)
           expect(result.successful).toBe(true)
-          expect(Connection.mockSend).nthCalledWith(1, getMsg(...media))
+          expect(Connection.mockSend).nthCalledWith(1, getMsg(media))
           done()
         })
         session.calling.notificationHandler(_playNotification)
@@ -325,7 +325,29 @@ describe('Call', () => {
         const action = await call.playAsync(...media)
         expect(action).toBeInstanceOf(PlayAction)
         expect(action.completed).toBe(false)
-        expect(Connection.mockSend).nthCalledWith(1, getMsg(...media))
+        expect(Connection.mockSend).nthCalledWith(1, getMsg(media))
+        session.calling.notificationHandler(_playNotification)
+        expect(action.completed).toBe(true)
+        done()
+      })
+
+      it('.play() with the new signature should wait until the playing ends', done => {
+        const params = { media, volume: 6.3 }
+        call.play(params).then(result => {
+          expect(result).toBeInstanceOf(PlayResult)
+          expect(result.successful).toBe(true)
+          expect(Connection.mockSend).nthCalledWith(1, getMsg(media, 6.3))
+          done()
+        })
+        session.calling.notificationHandler(_playNotification)
+      })
+
+      it('.playAsync() with the new signature should return a PlayAction for async control', async done => {
+        const params = { media, volume: 6.3 }
+        const action = await call.playAsync(params)
+        expect(action).toBeInstanceOf(PlayAction)
+        expect(action.completed).toBe(false)
+        expect(Connection.mockSend).nthCalledWith(1, getMsg(media, 6.3))
         session.calling.notificationHandler(_playNotification)
         expect(action.completed).toBe(true)
         done()
@@ -335,7 +357,7 @@ describe('Call', () => {
         call.playAudio('audio.mp3').then(result => {
           expect(result).toBeInstanceOf(PlayResult)
           expect(result.successful).toBe(true)
-          expect(Connection.mockSend).nthCalledWith(1, getMsg(media[0]))
+          expect(Connection.mockSend).nthCalledWith(1, getMsg([media[0]]))
           done()
         })
         session.calling.notificationHandler(_playNotification)
@@ -345,7 +367,27 @@ describe('Call', () => {
         const action = await call.playAudioAsync('audio.mp3')
         expect(action).toBeInstanceOf(PlayAction)
         expect(action.completed).toBe(false)
-        expect(Connection.mockSend).nthCalledWith(1, getMsg(media[0]))
+        expect(Connection.mockSend).nthCalledWith(1, getMsg([media[0]]))
+        session.calling.notificationHandler(_playNotification)
+        expect(action.completed).toBe(true)
+        done()
+      })
+
+      it('.playAudio() with volume should wait until the playing ends', done => {
+        call.playAudio({ url: 'audio.mp3', volume: 5 }).then(result => {
+          expect(result).toBeInstanceOf(PlayResult)
+          expect(result.successful).toBe(true)
+          expect(Connection.mockSend).nthCalledWith(1, getMsg([media[0]], 5))
+          done()
+        })
+        session.calling.notificationHandler(_playNotification)
+      })
+
+      it('.playAudioAsync() with volume should return a PlayAction for async control', async done => {
+        const action = await call.playAudioAsync({ url: 'audio.mp3', volume: 5 })
+        expect(action).toBeInstanceOf(PlayAction)
+        expect(action.completed).toBe(false)
+        expect(Connection.mockSend).nthCalledWith(1, getMsg([media[0]], 5))
         session.calling.notificationHandler(_playNotification)
         expect(action.completed).toBe(true)
         done()
@@ -355,7 +397,7 @@ describe('Call', () => {
         call.playTTS({ text: 'hello jest' }).then(result => {
           expect(result).toBeInstanceOf(PlayResult)
           expect(result.successful).toBe(true)
-          expect(Connection.mockSend).nthCalledWith(1, getMsg(media[1]))
+          expect(Connection.mockSend).nthCalledWith(1, getMsg([media[1]]))
           done()
         })
         session.calling.notificationHandler(_playNotification)
@@ -365,7 +407,27 @@ describe('Call', () => {
         const action = await call.playTTSAsync({ text: 'hello jest' })
         expect(action).toBeInstanceOf(PlayAction)
         expect(action.completed).toBe(false)
-        expect(Connection.mockSend).nthCalledWith(1, getMsg(media[1]))
+        expect(Connection.mockSend).nthCalledWith(1, getMsg([media[1]]))
+        session.calling.notificationHandler(_playNotification)
+        expect(action.completed).toBe(true)
+        done()
+      })
+
+      it('.playTTS() with volume should wait until the playing ends', done => {
+        call.playTTS({ text: 'hello jest', volume: 4 }).then(result => {
+          expect(result).toBeInstanceOf(PlayResult)
+          expect(result.successful).toBe(true)
+          expect(Connection.mockSend).nthCalledWith(1, getMsg([media[1]], 4))
+          done()
+        })
+        session.calling.notificationHandler(_playNotification)
+      })
+
+      it('.playTTSAsync() with volume should return a PlayAction for async control', async done => {
+        const action = await call.playTTSAsync({ text: 'hello jest', volume: 4 })
+        expect(action).toBeInstanceOf(PlayAction)
+        expect(action.completed).toBe(false)
+        expect(Connection.mockSend).nthCalledWith(1, getMsg([media[1]], 4))
         session.calling.notificationHandler(_playNotification)
         expect(action.completed).toBe(true)
         done()
@@ -375,7 +437,7 @@ describe('Call', () => {
         call.playSilence(5).then(result => {
           expect(result).toBeInstanceOf(PlayResult)
           expect(result.successful).toBe(true)
-          expect(Connection.mockSend).nthCalledWith(1, getMsg({ type: 'silence', params: { duration: 5 } }))
+          expect(Connection.mockSend).nthCalledWith(1, getMsg([{ type: 'silence', params: { duration: 5 } }]))
           done()
         })
         session.calling.notificationHandler(_playNotification)
@@ -385,7 +447,7 @@ describe('Call', () => {
         const action = await call.playSilenceAsync(5)
         expect(action).toBeInstanceOf(PlayAction)
         expect(action.completed).toBe(false)
-        expect(Connection.mockSend).nthCalledWith(1, getMsg({ type: 'silence', params: { duration: 5 } }))
+        expect(Connection.mockSend).nthCalledWith(1, getMsg([{ type: 'silence', params: { duration: 5 } }]))
         session.calling.notificationHandler(_playNotification)
         expect(action.completed).toBe(true)
         done()
@@ -398,11 +460,17 @@ describe('Call', () => {
       const audio = { type: 'audio', params: { url: 'audio.mp3' } }
       const tts = { type: 'tts', params: { text: 'hello jest' } }
 
-      const getMsg = (...play: ICallingPlay[]) => new Execute({
-        protocol: 'signalwire_service_random_uuid',
-        method: 'calling.play_and_collect',
-        params: { node_id: call.nodeId, call_id: call.id, control_id: 'mocked-uuid', collect, play }
-      })
+      const getMsg = (media: ICallingPlay, volume: number = 0) => {
+        const params: any = { node_id: call.nodeId, call_id: call.id, control_id: 'mocked-uuid', collect, play: [media] }
+        if (volume != 0) {
+          params.volume = volume
+        }
+        return new Execute({
+          protocol: 'signalwire_service_random_uuid',
+          method: 'calling.play_and_collect',
+          params
+        })
+      }
 
       it('.prompt() should wait until the playing ends', done => {
         call.prompt(collect, audio).then(result => {
@@ -411,6 +479,19 @@ describe('Call', () => {
           expect(result.terminator).toEqual('#')
           expect(result.result).toEqual('12345')
           expect(Connection.mockSend).nthCalledWith(1, getMsg(audio))
+          done()
+        })
+        session.calling.notificationHandler(_collectNotification)
+      })
+
+      it('.prompt() with volume property should wait until the playing ends', done => {
+        let collect = { volume: -4, initial_timeout: 10, digits_max: 5, digits_terminators: '#', digits_timeout: 10 }
+        call.prompt(collect, audio).then(result => {
+          expect(result).toBeInstanceOf(PromptResult)
+          expect(result.successful).toBe(true)
+          expect(result.terminator).toEqual('#')
+          expect(result.result).toEqual('12345')
+          expect(Connection.mockSend).nthCalledWith(1, getMsg(audio, -4))
           done()
         })
         session.calling.notificationHandler(_collectNotification)
@@ -448,6 +529,17 @@ describe('Call', () => {
         done()
       })
 
+      it('.promptAudioAsync() with volume  should return a PromptAction for async control', async done => {
+        let collect = { volume: 6.7, initial_timeout: 10, digits_max: 5, digits_terminators: '#', digits_timeout: 10 }
+        const action = await call.promptAudioAsync(collect, 'audio.mp3')
+        expect(action).toBeInstanceOf(PromptAction)
+        expect(action.completed).toBe(false)
+        expect(Connection.mockSend).nthCalledWith(1, getMsg(audio, 6.7))
+        session.calling.notificationHandler(_collectNotification)
+        expect(action.completed).toBe(true)
+        done()
+      })
+
       it('.promptTTS() should wait until the collect finished', done => {
         call.promptTTS(collect, { text: 'hello jest' }).then(result => {
           expect(result).toBeInstanceOf(PromptResult)
@@ -455,6 +547,19 @@ describe('Call', () => {
           expect(result.terminator).toEqual('#')
           expect(result.result).toEqual('12345')
           expect(Connection.mockSend).nthCalledWith(1, getMsg(tts))
+          done()
+        })
+        session.calling.notificationHandler(_collectNotification)
+      })
+
+      it('.promptTTS() with volume should wait until the collect finished', done => {
+        let collect = { volume: 6.7, initial_timeout: 10, digits_max: 5, digits_terminators: '#', digits_timeout: 10 }
+        call.promptTTS(collect, { text: 'hello jest' }).then(result => {
+          expect(result).toBeInstanceOf(PromptResult)
+          expect(result.successful).toBe(true)
+          expect(result.terminator).toEqual('#')
+          expect(result.result).toEqual('12345')
+          expect(Connection.mockSend).nthCalledWith(1, getMsg(tts, 6.7))
           done()
         })
         session.calling.notificationHandler(_collectNotification)