+++ packages/node/CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - Add `url` property to `RecordAction`.
 - Add methods to `pause` and `resume` a PlayAction.
+- Ability to set volume playback on `play` and `prompt` methods, or through the asynchronous `PlayAction` and `PromptAction` objects.
 
 ## [2.2.0] - 2019-09-09
 ### Fixed